# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "List Number TV2 Resolver",
    "version": "0.3.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/resolve": {
      "get": {
        "summary": "Resolve",
        "operationId": "resolve_v1_resolve_get",
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Input"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prefix": {
      "get": {
        "summary": "Prefix",
        "operationId": "prefix_v1_prefix_get",
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Input"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefixResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Button": {
        "properties": {
          "slot": {
            "type": "integer",
            "title": "Slot"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "resource_key": {
            "type": "string",
            "title": "Resource Key"
          }
        },
        "type": "object",
        "required": [
          "slot",
          "resource_key"
        ],
        "title": "Button"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Listing": {
        "properties": {
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "list_number": {
            "type": "string",
            "title": "List Number"
          },
          "tldx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tldx"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "template_key": {
            "type": "string",
            "title": "Template Key"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/Resource"
            },
            "type": "array",
            "title": "Resources"
          },
          "buttons": {
            "items": {
              "$ref": "#/components/schemas/Button"
            },
            "type": "array",
            "title": "Buttons"
          }
        },
        "type": "object",
        "required": [
          "country_code",
          "list_number",
          "tldx",
          "display_name",
          "entity_type",
          "template_key"
        ],
        "title": "Listing"
      },
      "Match": {
        "properties": {
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "list_number": {
            "type": "string",
            "title": "List Number"
          },
          "tldx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tldx"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "qualified_input": {
            "type": "string",
            "title": "Qualified Input"
          },
          "exact": {
            "type": "boolean",
            "title": "Exact"
          }
        },
        "type": "object",
        "required": [
          "country_code",
          "list_number",
          "tldx",
          "display_name",
          "entity_type",
          "qualified_input",
          "exact"
        ],
        "title": "Match"
      },
      "PrefixResponse": {
        "properties": {
          "country_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Context"
          },
          "input": {
            "type": "string",
            "title": "Input"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/Match"
            },
            "type": "array",
            "title": "Matches"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "country_context",
          "input",
          "matches",
          "limit",
          "has_more"
        ],
        "title": "PrefixResponse"
      },
      "ResolveResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "country_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Context"
          },
          "input": {
            "type": "string",
            "title": "Input"
          },
          "listing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Listing"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "country_context",
          "input"
        ],
        "title": "ResolveResponse"
      },
      "Resource": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uri"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "key",
          "type"
        ],
        "title": "Resource"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}